AugmentedMatrix.solve() returns a float array instead of a Fraction array

+++ augmented_matrix/augmented_matrix.py
@@ -148,7 +148,7 @@ class AugmentedMatrix:
         self.partial_pivot()
         self.reduce_echelon()
         self.check_valid_solution()
-        return self._matrix
+        return self._matrix.astype(float)
 
 
 __all__ = ['AugmentedMatrix', 'NoSolutionError']
